+++ docs/changes.rst
@@ -10,6 +10,7 @@ Weblate 5.16.2
 .. rubric:: Bug fixes
 
 * Avoid displaying confusing status icons for ghost languages on project or category level.
+* Fixed missing plural source strings when creating new bilingual plural units.
 
 .. rubric:: Compatibility